% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-summary-row.R
\name{bind_summary_row}
\alias{bind_summary_row}
\title{Bind one or many summary rows to a data frame}
\usage{
bind_summary_row(data, cols, fns, labels_to = NULL, by = NULL, below = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{cols}{A tidy-select selection of columns to summarize.}

\item{fns}{A named list of functions, where the name gives the label in the
column determined by \code{labels_to}.}

\item{labels_to}{A character vector of length one specifying the column to
(create if necessary and) put the labels in.}

\item{by}{A tidy-select selection of columns to summarize by.}

\item{below}{Default is \code{TRUE}. Binds total rows below the data. If \code{FALSE},
binds total rows above the data.}
}
\value{
A data frame.
}
\description{
\code{bind_summary_row()} handles the simple case of adding a total row to a data
frame, but with much greater flexibility. Specify one or many summary rows
using your own custom summary functions with \code{fns}, control placement of
labels with \code{labels_to} and \code{below}, or add a summary row for each group with
\code{by}.
}
\examples{
library(dplyr)
data("starwars")

starwars |>
  select(name, height, mass) |>
  bind_summary_row(
    cols = c(height, mass),
    fns = list(
      "Total" = \(x) sum(x, na.rm = TRUE)
    ),
    labels_to = "name",
    below = FALSE
  )

starwars |>
  select(sex, name, height, mass) |>
  bind_summary_row(
    cols = c(height, mass),
    fns = list(
      "Maximum" = \(x) max(x, na.rm = TRUE),
      "Total" = \(x) sum(x, na.rm = TRUE)
    ),
    labels_to = " ",
    by = sex,
    below = FALSE
  )
}
