% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect_at}
\alias{inspect_at}
\alias{inspect_where}
\title{Inspect a row of a data frame in the global environment.}
\usage{
inspect_at(.data, ...)

inspect_where(.data, ...)
}
\arguments{
\item{.data}{\code{.data} argument from \code{dplyr::filter()} or \code{dplyr::slice()}.}

\item{...}{data-masking arguments from \code{dplyr::filter()} or
\code{dplyr::slice()}.}
}
\value{
\code{.GlobalEnv}
}
\description{
\code{inspect_at()} and \code{inspect_where()} are designed to help debug calls to
\code{mutate_rowwise()}. Send the contents of a single data frame row to the
global environment for inspection, choosing the row using either
\code{dplyr::slice()} syntax (\code{inspect_at()}) or \code{dplyr::filter()} syntax
(\code{inspect_where()}).
}
\examples{
library(dplyr)

data("starwars")

starwars |>
  inspect_at(1)

starwars |>
  inspect_where(name == "Darth Vader")
}
