% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-rowwise.R
\name{mutate_rowwise}
\alias{mutate_rowwise}
\title{Mutate (rowwise)}
\usage{
mutate_rowwise(.data, ...)
}
\arguments{
\item{.data}{\code{.data} argument from \code{dplyr::mutate()}.}

\item{...}{data-masking name-value pairs from \code{dplyr::mutate()}.}
}
\value{
An (ungrouped) object of the same type as \code{.data}.
}
\description{
A shortcut to save typing. Performs mutate operations "rowwise" and
\code{ungroup()}s the data frame afterwards.
}
\examples{
library(dplyr)
library(tidyr)

data("starwars")

starwars |>
  tidyr::nest(.by = sex) |>
  mutate_rowwise(
    model = list(
      lm(mass ~ height, data)
    )
  )
}
