% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-all-sheets.R
\name{read_all_sheets}
\alias{read_all_sheets}
\title{Read all the sheets in an excel file}
\usage{
read_all_sheets(
  path,
  .names = "sheet_names",
  range = list(NULL),
  col_names = TRUE,
  col_types = list(NULL),
  na = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  progress = readxl::readxl_progress(),
  .name_repair = "unique"
)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{.names}{A character vector of names to assign to the resulting list
of tables, or a function to apply to the workbook's sheet names. Default
is \code{readxl::excel_sheets(path)}}

\item{range}{As in \code{readxl::read_excel()}, but optionally a vector the same
size as the number of sheets in \code{path}.}

\item{col_names}{As in \code{readxl::read_excel()}, but optionally a vector the
same size as the number of sheets in \code{path}.}

\item{col_types}{As in \code{readxl::read_excel()}, but optionally a vector the
same size as the number of sheets in \code{path}.}

\item{na}{As in \code{readxl::read_excel()}, but optionally a vector the same
size as the number of sheets in \code{path}.}

\item{trim_ws}{As in \code{readxl::read_excel()}, but optionally a vector the same
size as the number of sheets in \code{path}.}

\item{skip}{As in \code{readxl::read_excel()}, but optionally a vector the same
size as the number of sheets in \code{path}.}

\item{n_max}{As in \code{readxl::read_excel()}, but optionally a vector the same
size as the number of sheets in \code{path}.}

\item{guess_max}{As in \code{readxl::read_excel()}, but optionally a vector the
same size as the number of sheets in \code{path}.}

\item{progress}{As in \code{readxl::read_excel()}, but optionally a vector the
same size as the number of sheets in \code{path}.}

\item{.name_repair}{As in \code{readxl::read_excel()}, but optionally a vector the
same size as the number of sheets in \code{path}.}
}
\value{
A named list of \code{data.frame}s
}
\description{
A wrapper around \code{readxl::read_excel()} that reads all the sheets in an
excel file.
}
\examples{
\dontrun{
read_all_sheets("my-workbook.xlsx")
}
}
