% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-workbook.R
\name{write_workbook}
\alias{write_workbook}
\title{Write one or many tables to an excel workbook, with optional NERA-styling}
\usage{
write_workbook(
  tables,
  file_name,
  sheet_names = names(tables),
  headers = c("Privileged and Confidential", "Prepared at the Request of Counsel"),
  font_name = "Times New Roman",
  font_size = 11,
  bold_column_names = TRUE,
  add_filters = TRUE,
  auto_fit = TRUE,
  min_width = 8.43,
  max_width = 25.29,
  freeze_panes = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{tables}{A list of \code{data.frame}s or a single \code{data.frame}.}

\item{file_name}{The name of the file to write to.}

\item{sheet_names}{A character vector of sheet names, or a function to apply
to names(tables)}

\item{headers}{A character vector, one element for each header line. Use
\code{NULL} for no headers.}

\item{font_name}{A font name to use as the default in the workbook.}

\item{font_size}{A font size to use as the default in the workbook.}

\item{bold_column_names}{If \code{TRUE} column names are bold.}

\item{add_filters}{If \code{TRUE} adds filters to columns.}

\item{auto_fit}{If \code{TRUE} auto-fits columns (imperfectly).}

\item{min_width}{Minimum column width for auto-fitting.}

\item{max_width}{Maximum column width for auto-fitting.}

\item{freeze_panes}{If \code{TRUE} freezes table at row below column names.}

\item{overwrite}{If \code{TRUE} overwrites file.}
}
\description{
\code{write_workbook()} allows you to write a list of tables to different sheets
of the same excel workbook via a single function call. Use additional
arguments to control formatting details and add headers.
}
\examples{
df1 <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6))
df2 <- data.frame(C = c(1, 2, 3), D = c(4, 5, 6))
\dontrun{
write_workbook(
  list(df1, df2),
  c("Table 1", "Table 2"),
  "my-workbook"
)
}
}
